requirements.txt
appium>=2.0.0
Appium-Python-Client>=3.0.0
selenium>=4.0.0
pytest>=7.0.0
pytest-xdist>=3.0.0
pytest-rerunfailures>=11.0
pytest-timeout>=2.1.0
pytest-allure-adaptor>=1.11.0
allure-pytest>=2.13.0
allure-python-commons>=2.13.0
PyYAML>=6.0
python-dateutil>=2.8.0
colorlog>=6.7.0
pillow>=9.0.0